Refactored the function get_imprinting_probabilities().
Previously, the function was hard-coded to calculate influenza A imprinting, so e.g. there were only four possible imprinting types: H1N1, H2N2, H3N2 and naive.

With these modifications, the function is much more flexible. It loads subtype-specific circulation fractions as a data frame, `circulation_fractions`, with columns year, A/H1N1, A/H2N2, and A/H3N2.

These changes set up the function so that we will eventually be able to pass it data frames containing circulation fractions for arbitrary types of imprinting. The only requirement is a circulation_fractions data frame that contains a column for years from 1918:max_year, and an arbitrary number of columns containing influenza types, subtypes, strains, or clades that circulated in those years.